+++ src/pages/AddStartupPage.vue
@@ -1,23 +1,13 @@
 <template>
-
-  <q-page class="q-pa-md flex flex-center">
-    <div class="row justify-between items-center q-mx-auto full-width q-mt-lg" >
-      <div>    
-        <q-btn
-            flat
-            color="primary"
-            icon="arrow_back"
-            label=""
-            class=""
-            to="/more"
-          />
-        </div>
-        <div class="text-h6">
-          Add a Startup
-        </div>
-        <div class="q-px-lg"></div>
+  <q-page class="q-pa-md column justify-start">
+    <div class="row justify-between items-center q-mx-auto full-width q-mt-lg q-mb-xl">
+      <div>
+        <q-btn flat color="primary" icon="arrow_back" label="" class="" to="/more" />
+      </div>
+      <div class="text-h6">Add a Startup</div>
+      <div class="q-px-lg"></div>
     </div>
-      
+
     <div class="q-mx-auto" style="max-width: 500px; width: 100%">
       <AddStartupForm @startup-added="handleStartupAdded" />
     </div>